:sparkles: feat: add protected middleware for access control
- add logout endpoint
- add session status endpoint for access token validation
