+++ routes/auth-routes.mjs
@@ -1,13 +1,15 @@
 import express from "express";
 import { loginLimiter, registerLimiter } from "../middleware/limitHandler.mjs";
-import { accountInfo, login, register, verifyEmail } from '../controllers/auth-controller.mjs'
-import { protect } from "../middleware/authorization.mjs";
+import { accountInfo, login, logout, register, verifyEmail, checkSession } from '../controllers/auth-controller.mjs'
+import { protect } from "../middleware/authHandler.mjs";
 
 const router = express.Router();
 
 router.post('/register', registerLimiter, register);
 router.post('/login', loginLimiter, login);
+router.get('/logout', logout);
 router.post('/verify-email', verifyEmail); // ! Add limiter?
 router.get('/account-info', protect, accountInfo);
+router.get('/session-status', protect, checkSession);
 
 export default router;
